changing schedule information

+++ usefulprograms/factors.c
@@ -2,20 +2,20 @@
 #include<math.h>
 
 int main(int argc, char *argv){
-	for (int i = 1; i < pow(2, 63)-1; i*=6) {
-		int num = i;
-		int factors[num];
-		int numOfFactors = 0;
-		int cnt = 0;
-		for (long long int x = 0; x < num; x++){
-			if (((double)num)/((double)x) == (double)((int)((double)num/(double)x))){
+	for (unsigned int i = 1; i < pow(2, 63)-1; i*=6) {
+		unsigned int num = i;
+		unsigned int factors[num];
+		unsigned int numOfFactors = 0;
+		unsigned int cnt = 0;
+		for (unsigned int x = 0; x < num; x++){
+			if (((double)num)/((double)x) == (double)((unsigned int)((double)num/(double)x))){
 				factors[cnt++] = num/x;
 				numOfFactors++;
 			}
 		}
 		factors[cnt] = 1;
 		numOfFactors++;
-		printf("%d    Number of factors: %d\n", num, numOfFactors);
+		printf("%u    Number of factors: %u\n", num, numOfFactors);
 	}
 	return 0;
 }
